feat: 쉼표를 포함한 문자열을 분리해서 계산한다
