+++ README.md
@@ -5,7 +5,7 @@
 * 문자열 계산기에 숫자 이외의 값 또는 음수를 전달하는 경우 RuntimeException 예외를 throw 한다.
 
 ## 기능 목록
-- [ ] 쉼표를 포함한 문자열을 분리해서 계산한다.
+- [X] 쉼표를 포함한 문자열을 분리해서 계산한다.
 - [ ] 콜론을 포함한 문자열을 분리해서 계산한다.
 - [ ] 쉼표와 콜론을 포함한 문자열을 분리해서 계산한다.
 - [ ] 빈 문자열 또는 null을 입력할 경우 0을 반환해야 한다.
