+++ src/main/java/pluscalculator/domain/PlusCalculator.java
@@ -1,0 +1,12 @@
+package pluscalculator.domain;
+
+import java.util.Arrays;
+import java.util.stream.Collectors;
+
+public class PlusCalculator {
+
+    public static Numbers extractNumbers(final String plusExpression) {
+        String[] splitNumbers = plusExpression.split(",");
+        return new Numbers(Arrays.stream(splitNumbers).map(Integer::parseInt).collect(Collectors.toList()));
+    }
+}
